refactor: move unit tests to mod file
